implement bottom nav
*add home screen
*add checklist screen
*add settings screen
