+++ Orchen/Orchen/Controller/CheckListView.swift
@@ -1,5 +1,5 @@
 //
-//  HomeView.swift
+//  CheckListView.swift
 //  Orchen
 //
 //  Created by Rahul Ramjeawon on 11/14/24.
@@ -8,11 +8,12 @@
 
 import SwiftUI
 
-struct HomeView: View {
+struct CheckListView: View {
     var body: some View {
         TabView {
             ZStack {
-                Text("hello this is the home view")
+                BackgoundColor()
+                Text("hello this is the check list view")
             }
         }
     }
