+++ Orchen/Orchen/Controller/SettingView.swift
@@ -1,5 +1,5 @@
 //
-//  HomeView.swift
+//  SettingView.swift
 //  Orchen
 //
 //  Created by Rahul Ramjeawon on 11/14/24.
@@ -8,11 +8,11 @@
 
 import SwiftUI
 
-struct HomeView: View {
+struct SettingView: View {
     var body: some View {
         TabView {
             ZStack {
-                Text("hello this is the home view")
+                Text("hello this is the setting view")
             }
         }
     }
