+++ Orchen/Orchen/GoogleService-Info.plist
@@ -1,1 +1,30 @@
-
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyDmR8oWTHvdW6xVheX-bCYSeu39s5GVnl4</string>
+	<key>GCM_SENDER_ID</key>
+	<string>415813335098</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.orchen.media.Orchen</string>
+	<key>PROJECT_ID</key>
+	<string>orchen-90a36</string>
+	<key>STORAGE_BUCKET</key>
+	<string>orchen-90a36.firebasestorage.app</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:415813335098:ios:c17c91eeea19ec3d605d03</string>
+</dict>
+</plist>
